refactor: rename lotties folder
