+++ src/assets/animations/loading.json
