+++ src/assets/animations/user-details-skeleton.json
